requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.9
requests==2.24.0
robotframework==3.2.1
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.9
